+++ Assets/Scripts/Main Scripts/Tile_Data.cs
@@ -18,7 +18,7 @@ public class Tile_Data : ScriptableObject
     [SerializeField] public float Average_Heat_Level; //Global Warming, Events and Neutral Disasters effect heat levels. Heat levels directly effects Habitability level. 
 
     //Pop
-    [SerializeField] public int Population;
+    [SerializeField] public float Population;
 
     //Tree Type and Growth
     [SerializeField] public Tree Tree_Plantation;
